Make 4 byte ngrams work.
Cast the 1 constant to a 64 bit value so the bitshifting works properly
in the case of a 4 byte ngram.

+++ src/fileIndexer.cpp
@@ -50,7 +50,7 @@ filesProcessed(0), verbose(true)
 
     if(ngramLength == 3  || ngramLength == 4){
         this->ngramLength = ngramLength;
-        this->maxNgram = (1 << (8*ngramLength));
+        this->maxNgram = ((uint64_t) (1) << (8*ngramLength));
         this->pagesize = getpagesize();
     }
     else {
